% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalsens.R
\docType{data}
\name{lalonde.psid}
\alias{lalonde.psid}
\title{Non-experimental data from Lalonde (1986)}
\format{A data frame with 2675 rows and 12 variables}
\usage{
data(lalonde.psid)
}
\description{
A dataset of experimental treated units and non-experimental
control units from the Panel Study of Income Dynamics (PSID).
}
\details{
\itemize{
  \item \code{age} - age in years.
  \item \code{education} - number of years of schooling.
  \item \code{black} - 1 if black, 0 otherwise.
  \item \code{hispanic} - 1 if Hispanic, 0 otherwise.
  \item \code{married} - 1 if married, 0 otherwise.
  \item \code{nodegree} - 1 if no high school degree, 0 otherwise.
  \item \code{re74} - earnings ($) in 1974.
  \item \code{re75} - earnings ($) in 1975.
  \item \code{re78} - earnings ($) in 1978.
  \item \code{u74} - 1 if unemployed in 1974, 0 otherwise.
  \item \code{u75} - 1 if unemployed in 1975, 0 otherwise.
  \item \code{treat} - 1 if treated, 0 otherwise.
}
}
\references{
LaLonde, Robert J. (1986). Evaluating the Econometric
Evaluations of Training Programs with Experimental Data. The
American Economic Review, 76(4), 604--620.
}
\keyword{datasets}
