% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causalsens.R
\name{one.sided}
\alias{one.sided}
\alias{alignment}
\alias{one.sided.att}
\alias{alignment.att}
\title{Confounding functions}
\usage{
one.sided(alpha, pscores, treat)

alignment(alpha, pscores, treat)

one.sided.att(alpha, pscores, treat)

alignment.att(alpha, pscores, treat)
}
\arguments{
\item{alpha}{vector of confounding values to use in the
sensitivity analysis.}

\item{pscores}{vector of propensity scores for each unit.}

\item{treat}{vector of treatment values for each unit.}
}
\description{
Various confounding functions for use with
\code{\link{causalsens}}.
}
